api-gateway/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
cryptography==44.0.1
ecdsa==0.19.0
fastapi==0.115.9
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
PyJWT==2.10.1
python-dotenv==1.0.1
python-jose==3.4.0
redis==5.2.1
rfc3986==1.5.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
typing_extensions==4.12.2
upgrade-requirements==1.7.0
uvicorn==0.34.0
